+++ CHANGELOG.md
@@ -22,6 +22,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for Oracle database.
 - Make into Java 9 module.
 
+## [Unreleased]
+### Added
+- Result now keeps the Column info alongside each column mapping for access to 
+  the column metadata when accessing the result.
+
 ## [0.3.9] - 2021-04-08
 ### Added
 - `bin` macro function which return the interval that a value belongs to and